move `IntoHash` implementation outside of hash.rs
